36989 common code in integration algos moved to a util module
